+++ app/src/main/java/com/example/ztoreme_1/productos/ActivityActualizar.kt
@@ -126,8 +126,8 @@ class ActivityActualizar : AppCompatActivity() {
                     (editNumCantidad.text.toString()).toInt(),
                     (editstockMax.text.toString()).toInt(),
                     (editstockMin.text.toString()).toInt(),
-                    (editprecioCompra.text.toString()).toInt(),
-                    (editprecioVenta.text.toString()).toInt()
+                    (editprecioCompra.text.toString()).toDouble(),
+                    (editprecioVenta.text.toString()).toDouble()
                 )
